// app/model/parametro.ts
export class Parametro {

  id: number;
  dataInizio: string;
  dataFine: string;
  nome: string;
  nomeCampoRif: string;
  unitaMisura: string;
  valoreMax: number;
  valoreMin: number;
  tipo: string;
  userCreate: string;
  userModify: string;
  tmstCreate: string;
  tmstModify: string;
  valoreDefault: string;


  constructor() {
    this.id = null;
    this.dataInizio = null;
    this.dataFine = null;
    this.nome = null;
    this.nomeCampoRif = null;
    this.unitaMisura = null;
    this.valoreMax = null;
    this.valoreMin = null;
    this.tipo = null;
    this.userCreate = null;
    this.userModify = null;
    this.tmstCreate = null;
    this.tmstModify = null;
    this.valoreDefault = null;
  }


}